refactor: 프로필 페이지 컴포넌트 폴더 구조 변경
프로필 페이지 관련 컴포넌트들을 기능별 폴더로 재구성하여
코드 관리 및 유지보수성을 향상시켰습니다.
- modals 폴더에 있던 모달 컴포넌트들을 각 컴포넌트 관련 폴더로 이동
- 스타일 파일들의 위치도 변경하여 컴포넌트와 함께 관리
