+++ src/pages/ProfilePage/components/edit-nickname-modal/styles.ts
